added check that folder exists before trying to read files within

+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress.config.ts
@@ -48,16 +48,21 @@ module.exports = defineConfig({
 
         // Custom task to check if files exist in the downloads folder
         checkForFiles(folderPath) {
-          const files = fs.readdirSync(folderPath);
-          return files.length > 0 ? files : null;
+          if (fs.existsSync(folderPath)) {
+            const files = fs.readdirSync(folderPath);
+            return files.length > 0 ? files : null;
+          }
+          return null;
         },
 
         // Custom task to delete all files in the downloads folder
         clearDownloads(folderPath) {
-          const files = fs.readdirSync(folderPath);
-          files.forEach((file) => {
-            fs.unlinkSync(path.join(folderPath, file));
-          });
+          if (fs.existsSync(folderPath)) {
+            const files = fs.readdirSync(folderPath);
+            files.forEach((file) => {
+              fs.unlinkSync(path.join(folderPath, file));
+            });
+          }
           return { success: true };
         }
       });
